app/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-auth==2.38.0
google-cloud-aiplatform==1.86.0
pydantic==2.6.1
google-genai==1.13.0